requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
python-telegram-bot==20.8
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.1.4
pytz==2024.1